Reduce duplicate work in building PHP images, make matrix generation dynamic

+++ php/scripts/build.sh
@@ -37,14 +37,18 @@ fi
 ## latest version; build matrix will override to build each supported version
 VERSION_LIST="${VERSION_LIST:-"7.4"}"
 VARIANT_LIST="${VARIANT_LIST:-"cli cli-loaders fpm fpm-loaders"}"
+MINOR_VERSION=""
+BUILT_TAGS=()
 
-docker buildx create --use
-IMAGE_NAME="${IMAGE_NAME:-"davidalger/php"}"
+##### docker buildx create --use
+IMAGE_NAME="${IMAGE_NAME:-"ghcr.io/wardenenv/centos-php"}"
 if [[ "${INDEV_FLAG:-1}" != "0" ]]; then
   IMAGE_NAME="${IMAGE_NAME}-indev"
 fi
 for BUILD_VERSION in ${VERSION_LIST}; do
-  MAJOR_VERSION="$(echo "${BUILD_VERSION}" | sed -E 's/([0-9])([0-9])/\1.\2/')"
+  MAJOR_VERSION="$(echo "${BUILD_VERSION}" | sed -E 's/^([0-9]+\.[0-9]+).*$/\1/')"
+  echo "### PHP ${MAJOR_VERSION} Tags" >> $GITHUB_STEP_SUMMARY
+
   for BUILD_VARIANT in ${VARIANT_LIST}; do
     # Configure build args specific to this image build
     export PHP_VERSION="${MAJOR_VERSION}"
@@ -54,14 +58,31 @@ for BUILD_VERSION in ${VERSION_LIST}; do
     printf "\e[01;31m==> building %s:%s (%s)\033[0m\n" \
       "${IMAGE_NAME}" "${BUILD_VERSION}" "${BUILD_VARIANT}"
 
-    docker build -t "${IMAGE_NAME}:build" "${BUILD_VARIANT}" $(printf -- "--build-arg %s " "${BUILD_ARGS[@]}")
+    # Build for all platforms at once
+    docker buildx build \
+      --platform=linux/amd64,linux/arm64 \
+      --cache-from=type=registry,ref=image:buildcache-amd64 \
+      --cache-to=type=registry,ref=image:buildcache-amd64 \
+      --cache-from=type=registry,ref=image:buildcache-arm64 \
+      --cache-to=type=registry,ref=image:buildcache-arm64 \
+      --tag "${IMAGE_NAME}:build" \
+      "${BUILD_VARIANT}" \
+      $(printf -- "--build-arg %s " "${BUILD_ARGS[@]}")
+    
+    # Load the built image to run it temporarily
+    docker buildx build --load \
+      --tag "${IMAGE_NAME}:build" \
+      "${BUILD_VARIANT}" \
+      $(printf -- "--build-arg %s " "${BUILD_ARGS[@]}")
 
     # Strip the term 'cli' from tag suffix as this is the default variant
     TAG_SUFFIX="$(echo "${BUILD_VARIANT}" | sed -E 's/^(cli$|cli-)//')"
     [[ ${TAG_SUFFIX} ]] && TAG_SUFFIX="-${TAG_SUFFIX}"
 
     # Fetch the precise php version from the built image and tag it
-    MINOR_VERSION="$(docker run --rm -t --entrypoint php "${IMAGE_NAME}:build" -r 'echo phpversion();')"
+    if [[ "${BUILD_VARIANT}" == "cli" ]]; then
+      MINOR_VERSION="$(docker run --rm -t --entrypoint php "${IMAGE_NAME}:build" -r 'echo phpversion();')"
+    fi
 
     # Generate array of tags for the image being built
     IMAGE_TAGS=(
@@ -69,9 +90,55 @@ for BUILD_VERSION in ${VERSION_LIST}; do
       "${IMAGE_NAME}:${MINOR_VERSION}${TAG_SUFFIX}"
     )
 
+    # Update the tags for the image, will use build cache
+    docker buildx build \
+      --platform=linux/arm64,linux/amd64 \
+      --cache-from=type=registry,ref=image:buildcache-amd64 \
+      --cache-to=type=registry,ref=image:buildcache-amd64 \
+      --cache-from=type=registry,ref=image:buildcache-arm64 \
+      --cache-to=type=registry,ref=image:buildcache-arm64 \
+      --tag "${IMAGE_NAME}:${MAJOR_VERSION}${TAG_SUFFIX}" \
+      --tag "${IMAGE_NAME}:${MINOR_VERSION}${TAG_SUFFIX}" \
+      "${BUILD_VARIANT}" \
+      $(printf -- "--build-arg %s " "${BUILD_ARGS[@]}")
+    
+    BUILT_TAGS+=("${IMAGE_TAGS[@]}")
+
+    # Load and push tagged images to cache
+    #    Separate because of https://github.com/docker/buildx/issues/59
+    #  These are alrady in the cache, so this step should only take a few seconds
+    docker buildx build --load \
+      --platform=linux/amd64 \
+      --cache-from=type=registry,ref=image:buildcache-amd64 \
+      --cache-to=type=registry,ref=image:buildcache-amd64 \
+      --tag "${IMAGE_NAME}:${MAJOR_VERSION}${TAG_SUFFIX}" \
+      --tag "${IMAGE_NAME}:${MINOR_VERSION}${TAG_SUFFIX}" \
+      "${BUILD_VARIANT}" \
+      $(printf -- "--build-arg %s " "${BUILD_ARGS[@]}")
+    
+    docker buildx build --load \
+      --platform=linux/arm64 \
+      --cache-from=type=registry,ref=image:buildcache-arm64 \
+      --cache-to=type=registry,ref=image:buildcache-arm64 \
+      --tag "${IMAGE_NAME}:${MAJOR_VERSION}${TAG_SUFFIX}" \
+      --tag "${IMAGE_NAME}:${MINOR_VERSION}${TAG_SUFFIX}" \
+      "${BUILD_VARIANT}" \
+      $(printf -- "--build-arg %s " "${BUILD_ARGS[@]}")
+
     # Iterate and push image tags to remote registry
-    if [[ ${PUSH_FLAG} != 0 ]]; then
-      docker buildx build --push --platform=linux/arm64,linux/amd64 -t "${IMAGE_NAME}:${MAJOR_VERSION}${TAG_SUFFIX}" -t "${IMAGE_NAME}:${MINOR_VERSION}${TAG_SUFFIX}" "${BUILD_VARIANT}" $(printf -- "--build-arg %s " "${BUILD_ARGS[@]}")
-    fi
+    # if [[ ${PUSH_FLAG} != 0 ]]; then
+    #   docker buildx build \
+    #     --push \
+    #     --platform=linux/arm64,linux/amd64 \
+    #     -t "${IMAGE_NAME}:${MAJOR_VERSION}${TAG_SUFFIX}" \
+    #     -t "${IMAGE_NAME}:${MINOR_VERSION}${TAG_SUFFIX}" \
+    #     "${BUILD_VARIANT}" \
+    #     $(printf -- "--build-arg %s " "${BUILD_ARGS[@]}")
+    # fi
   done
+
+  echo "$(jq -R 'split(" ")' <<< "${BUILT_TAGS[@]}")" >> $GITHUB_STEP_SUMMARY
+  echo "::notice title=PHP ${MAJOR_VERSION} Tags to Push::${BUILT_TAGS}" >> $GITHUB_OUTPUT
 done
+
+echo "built_tags=$(jq -cR 'split(" ")' <<< "${BUILT_TAGS[@]}")" >> $GITHUB_OUTPUT
